fix: 코드레빗 리뷰 반영
